+++ test/resources/unused_code_analyzer/public_members.dart
@@ -160,3 +160,26 @@ typedef Hello = String;
 
 // LINT
 typedef World = int;
+
+const double _kMenuVerticalPadding = 1;
+
+void setPadding() {
+  const padding = _kMenuVerticalPadding;
+}
+
+// LINT
+class MyOtherWidget extends StatefulWidget {
+  MyOtherWidget();
+
+  @override
+  _MyOtherWidgetState createState() => _MyOtherWidgetState();
+}
+
+class _MyOtherWidgetState extends BaseState<MyOtherWidget> {
+  @override
+  Widget build(BuildContext context) => ElevatedButton(
+        onPressed: () {},
+        onLongPress: () {},
+        child: Text('PRESS'),
+      );
+}